utils/procon_tools: devide source execution
