+++ utils/procon_tools/languages/__init__.py
@@ -1,1 +1,2 @@
-from .cpp_utils import compile_cpp_source
+from .cpp_utils import compile_cpp_source, exec_cpp_input
+from .python_utils import exec_py3_input
